Experimental refactoring to support multi-column output
The existing code ties us to one CTE per column in the study. We are
going to need to break this in cases where we want a measurement value
and the date on which that value was obtained. BMI is the immediate
example of this.

Another approach might be to allow a single column in the study
definition to produce several columns in the output. For instance,
where you had something like:

   ...
   bmi = patients.latest_bmi(include_date=True)
  ...

It would produce two columns in the output `bmi` and `bmi_date`.