Support returning multiple columns from a CTE
We give the first returned column the name of the desired output column
and then we treat any subsequent column names as suffixes. For example,
if a query returns ["measurement_value", "date_measured"] and we assign
it to the output column "bmi" it will produce two columns in the output:
"bmi" and "bmi_date_measured".

Previously each returned column would be given the same output name
which (surprisingly) didn't result in an error but did give "last one
wins" behaviour.