Add mean recorded value covariate
This is to support blood pressure readings and is a more general
replacement for #18

This does mean that we end up with separate date columns for the
systolic and diastolic readings and that we drop the requirement that
both readings need to come from the same day. However this matches the
original specification for the covariate and is unlikely to make any
difference in practice.