Change string literal handling for expressions
We now require they be quoted as normal rather than special-casing `M`
and `F`. This is to allow us to support a wider range of categorical
variables.

String literals can only contain alphanumerics and the underscore
character and can be a maximum of 16 characters long. They may be single
or double quoted in the source although they will be re-formatted using
the appopriate quoting convention for the backend.

This is a backwards incompatible change although the existing study is
not using the M/F behaviour at preseent so it shouldn't cause any
issues.