Support arbitrary logical expression for categorical variables
This extends the `satisfying` boolean predicate with a `categorised_as`
predicate which can handle an arbitrary number of categories. These are
expressed as a dictionary and matched in the supplied order. One
category can be defined with the expression "DEFAULT" making it the
fallback value if none of the other expressions match. If not default is
specified then the fallback value is the empty string.

For example:
```py
StudyDefinition(
    smoking_status=patients.categorised_as(
        {
            "S": "most_recent_smoking_code = 'S'",
            "E": """
                 most_recent_smoking_code = 'E' OR (
                   most_recent_smoking_code = 'N' AND ever_smoked
                 )
            """,
            "N": "most_recent_smoking_code = 'N' AND NOT ever_smoked",
            "M": "DEFAULT"
        },
        most_recent_smoking_code=patients.with_these_clinical_events(
            smoking_codes,
            find_last_match_in_period=True,
            on_or_before='2020-02-01',
            returning="category",
        ),
        ever_smoked=patients.with_these_clinical_events(
            filter_codes_by_category(smoking_codes, include=['S', 'E']),
            on_or_before='2020-02-01'
        ),
    )
)
```

We also add a ` filter_codes_by_category` function which takes a
codelist and a list of categories and returns a new codelist with just
the entries matching those categories e..g
```py
 filter_codes_by_category(smoking_codes, include=['S', 'E'])
```

A small breaking change is that we no longer accept unquoted `M` and `F`
as string literals - they must now be quoted in the normal fashion.