Fix collation mismatch issues
When we tried to run against the dummy data it would fail with the
error:
```
Cannot resolve the collation conflict between
"SQL_Latin1_General_CP1_CI_AS" and "Latin1_General_BIN" in the equal to
operation. (468) (SQLExecDirectW)
```

This is because Read 3 codes are case-senstive and so the CTV3Code
column has a case-senstive collation.