Inner join each query on the population table
This prevents us from e.g. calculating the BMI of every patient in the
system only to throw most of these results away later because our study
population is only a small subset of all patients.

How much performance difference this will make depends on what fraction
of the patients in the database are in the study. At some sufficiently
high ratio it will presumably reduce performance as it introduces an
extra join for each covariate that doesn't end up removing many rows.

With only the dummy data to test against it's hard to know what's best
so I'll leave this PR here in case it's needed.