Support expression queries in population definition
The major change here is that previously we always assumed that we would
have a `#population` table to use as a primary table to which everything
else was LEFT JOINed. Now we accept that we might not have such a table
(because `population` may just be an expression over other values) so in
this case we use the `Patient` table as our primary table.

In either case we add a WHERE condition to ensure that whatever value is
defined by `population` evalutes to 1.