Support query we need to count COPD exacerbations
This adds some extra options to the `with_these_clinical_events`
function:
```py
study = StudyDefinition(
    population=patients.all(),
    episode_count=patients.with_these_clinical_events(
        lrti_and_aecopd_codes,
        ignore_events_where_these_codes_occur_on_same_day=annual_review_codes,
        on_or_before="2020-02-01",
        returning="number_of_episodes_in_period",
        only_count_events_as_new_episode_when=">14 days after end of previous episode",
    ),
)
```

This gives the impression of being a slightly more general API than it
actually is. In particular, the definition for the washout period only
accepts exactly one format of string with the number of days allowed to
vary. The guiding principle I've followed is to keep the API explict and
as self-documenting as possible.

The SQL implementation itself uses the `LAG` function which, I believe,
gives us a fairly efficient way of getting "time since previous event".
Where this is less or equal to the washout period we set the
"is_new_episode" flag to 0, otherwise we set it to 1. Summing over these
flags gives us the episode count.

The "ignore on same day" clause is implemented using a correlated
subquery which may or may not have good enough performance in the real
data.