Automatically set category ratios for `satisfying`
The `satisfying` method is just syntatic sugar over `categorised_as`
with `1` and `0` as the two categories. We need to handle this fact in
the `return_expectations` argument as well.

(Note, this doesn't quite match the "what we want" example in the linked
issue because `earliest` is a required argument to `date` as well as
`latest`. I assume this was an accidental omission.)

Closes #106