Support querying SGSS tables
This allows us to check COVID test results in the Second Generation
Surveillance System. Example queries are:
```py
study = StudyDefinition(
    population=patients.all(),
    positive_covid_test_ever=patients.with_test_result_in_sgss(
        pathogen="SARS-CoV-2",
        test_result="positive"
    ),
    negative_covid_test_ever=patients.with_test_result_in_sgss(
        pathogen="SARS-CoV-2",
        test_result="negative"
    ),
    tested_for_covid=patients.with_test_result_in_sgss(
        pathogen="SARS-CoV-2",
        test_result="any",
        on_or_before="2020-05-01"
    ),
    first_positive_test_date=patients.with_test_result_in_sgss(
        pathogen="SARS-CoV-2",
        test_result="positive",
        find_first_match_in_period=True,
        returning="date",
        date_format="YYYY-MM-DD",
    ),
)
```

Only SARS-CoV-2 results are included in our data extract so this will
throw an error if the specified pathogen is anything other than
"SARS-CoV-2".

`test_result` must be one of: "positive", "negative" or "any"

The date used is the date the specimen was taken. Where a patient has
multiple positive results only the date of the earliest specimen that
tested positive is recorded.

The data we receive has an Organism_Species_Name field which is always
"SARS-CoV-2 CORONAVIRUS (Covid-19)" in the case of a positive result and
"NEGATIVE SARS-CoV-2 (COVID-19)" in the case of a negative result. The
code will throw an error if anything else is ever found in this field.