Support querying Care Homes table
TPP have attempted to match patient addresses to care homes as stored in
the CQC database. At its most simple the `care_home_status` query
returns a boolean indicating whether the patient's address (as of the
supplied time) matched with a care home.

It is also possible to return a more complex categorisation based on
attributes of the care homes in the CQC database, which can be freely
downloaded here:
https://www.cqc.org.uk/about-us/transparency/using-cqc-data

At present the only imported fields are:
    LocationRequiresNursing
    LocationDoesNotRequireNursing

But we can ask for more fields to be imported if needed.

The `categorised_as` argument acts in effectively the same way as for
the `categorised_as` function except that the only columns that can be
referred to are those belonging to the care home table (i.e. the two
nursing fields above) and the boolean `IsPotentialCareHome`.

Example queries:
```py
is_in_care_home=patients.care_home_status_as_of("2020-01-01")

care_home_type=patients.care_home_status_as_of(
    "2020-01-01",
    categorised_as={
        "PC": """
          IsPotentialCareHome
          AND LocationDoesNotRequireNursing='Y'
          AND LocationRequiresNursing='N'
        """,
        "PN": """
          IsPotentialCareHome
          AND LocationDoesNotRequireNursing='N'
          AND LocationRequiresNursing='Y'
        """,
        "PS": "IsPotentialCareHome",
        "U": "DEFAULT",
    },
)
```