Refactor type inference for columns
We want to be able to do this entirely separately from building the
queries, which involves code specific to a particular backend.

This does mean that adding a new query type involves changes in at least
two separate files: one to define how the query works on the backend and
another to define the resulting column type. However now that we move
towards multiple backends adding a new query type is necessarily a
multi-file affair anyway. And this will blow up straight away with a
reasonably comprehensible error message if you define a new query
without defining its type.

This involves a small breaking change, which I don't think affects any
of our studies: when defining categories using logical expressions it's
now mandatory to supply a default category (previously the empty string
was the default). I think this is good practice anyway so we may as well
require it.