Enforce a maximum as well as a minimum date of birth
It turns out there are some patients in the TPP database with dates of
birth in the future (sometimes by quite a long way). This is infrequent
enough that it's probably ignorable, but it would be better to exclude
them anyway.

At the same time we can exclude patients who are below the age of one
because the youngest patients we will use in our analysis are patients
who turn one this month, and on principle we shouldn't retreive data
we're not going to use.