feat: Support returning de-registrations dates
Normally we require that patients stay registered throughout the study
period, but in certain cases where we have data from outside primary
care we may want to include de-registered patients in a study. For such
patients we need to know the date they de-registered so we can treat
them appropriately in the analysis.