Refactor sqlcmd wrapper for better error handling
Previously this code used the `-o output_file` option. This caused both
query output and warnings/errors to be mixed together in a single file,
which then required hairy and error prone code to try to separate back
out.

We now just let sqlcmd write to stdout, which we redirect to a file
ourselves, leaving error output written to stderr.

Closes #239