feat: Support writing results to intermediate table
Sometimes the queries to extract study data will run successfully but
something will interrupt the download process and then the entire query
will need to be run again.

This PR adds support for writing results to a table in a temporary
database before downloading. If the download is interrupted then a
re-run of the extractaton process will pick up the previously created
table and download it without having to re-run the queries.

Once the data is successfully downloaded the table is deleted.

In order to use this feature the environment variable
`TEMP_DATABASE_NAME` must be set. This will require a change to the job
runner and updated instructions for those running extracts manually.

Fixes #241