Include hashed_organisation in every CREATE TABLE query
This is checked by the EMIS access guard to ensure that we are allowed
to access the data.  We never use it.

A patient's hashed_organisation corresponds to the practice they belong
to now, not the practice they belonged to at the time a recorded event
happened.

As such,

    GROUP BY registration_id, hashed_organisation

is equivalent to

    GROUP BY registration_id