feat: Distinguish partial and complete CSV files
When errors occur during `generate_cohort` we want to keep the output
files around for debugging purposes but we want it to be clear that they
are not complete.

This PR produces files named like `input.partial.20200901-154348.csv`
which are only renamed to `input.csv` once the download is complete and
checks are passed.

Closes #278