fix: Fix potential cache key clash in temporary tables
Our server contains both the full dataset and a smaller sample of the
data for rapid testing. It's not inconceivable that we'd end up running
the same extract on these at overlapping times. Previously this would
have caused a cache key collision, but now we include the name of the
database in the list of things which get hashed to create the key.