Support additional "setup" queries
As well as the queries we run to populate each column in the output, we
sometimes want to create intermediate, temporary tables to support these
queries. So far we've avoided doing this through use of CTEs,
sub-queries and the like. But this sometimes has big performance
implications.

We already support a specific case of intermediate tables, which are the
"codelist" tables, but we need something more general. This commit adds
support for arbitrary additional queries.