fix: Refactor "not on same day" query to use temp tables
This is based on the suggested SQL in #302, but implements only half of
it in the sense that it creates one indexed, temporary table rather than
two because this requires far less invasive refactoring of our existing
queries. My sense is that this will give us most of the speed up anyway
but we won't really know until we try it in production.