fix: Add ORDER BY clauses to make results deterministic
In certain cases where we use the `ROWNUMBER() OVER (PARTITION BY ..
ORDER BY .. )` pattern to select only the first matching row in a join,
we don't specify enough order conditions to uniquely determine a first
row. When this happens the result is arbitrary (which is sort of OK) and
non-deterministic (which is not OK). This patch attempts to fix the
non-determinism by adding a tie-breaker column to the end of the sort
conditions, which is usually the ID field of the table in question.

There are two cases (the EC and APCS tables) where we can't 100%
guarantee determinstic ordering because this are no fields with a unique
constraint. However in both cases there is an "ident" field which is
almost entirely unique and where the tiny handful of duplicates are
otherwise identical and so shouldn't result in non-deterministic outputs
anyway.

This ought to fix the issue identified in #297 although I haven't yet
confirmed this.

Note that while this fixes the non-determinism it doesn't address the
arbitrariness. That is, where we have multiple matches there might be
better reasons for choosing one result over another, or better ways of
combining those results together as in the case of BMI.