fix: Quick fix for inconsistent ONS death data
A patient appears twice in the ONS data with different dates of death
(offset by one day). This results in an error when running the extract.
As we need to extract this cohort urgently we're going to use the
minimum date for now, pending a full discussion of how best to handle
this kind of inconsistency.