Replace `make_date_filter` with `get_date_condition`
This latter function returns not just the SQL fragment with the
condition but also a (possibly empty) SQL fragment with JOINs that will
be required to evaluate the date condition. In order to do this it
requires a table name as an additional argument.

At present this only handles date literals so it doesn't do any more
than the previous function, but its API paves the way for further
changes.