fix: Escape characters rather than using a safelist
This is prompted by the need to query the High Cost Drugs data in which
drug names are (apparently) entered as free text. This leads to names
such as:

    infliximab-remicade!100mg!infusion
    infliximab 'remicade' (10ml vi
    infliximab 'remicade' (10ml vial)
    ixekizumabß

Given the presence of single quotes in these names the only way we can
use them in queries is to escape them properly.

We also take this opportunity to correctly escape wildcard characters in
strings which are going to be used in LIKE queries.