Use patient_no_duplicates instead of patient_all_orgs_v2
patient_all_orgs_v2 contains a handful of duplicate registration IDs.  This causes
problems: because of the way we build our final query with a series of N LEFT JOINs, a
registration ID appearing twice in the patient table can appear up to 2^N times.

EMIS have not yet fixed this, so for now I have created a table called
patient_no_duplicates which removes these duplicate registration IDs.

WE MUST REMEMBER TO RECREATE THIS TABLE EACH TIME THE DATA IS REFRESHED!