fix: Create temporary patient table on each run
The patient_all_orgs_v2 table contains a handful of duplicate
registration IDs.  Rather than having to remember to create a table with
duplicates remove on each database build, it makes more sense to build
this table from scratch on each run.

There are no new tests.  I have given up on trying to cajole SQLAlchemy
+ SQL Server into letting me create a scenario where there are
duplicate registration_ids.  Either registration_id is a primary key, in
which case it can't contain duplicates, or it's not, in which case
SQLAlchemy refuses to create a foreign key to it.  I don't see a way
around this without rewriting the tests that depend on foreign keys (ie
all of them).