fix: Output consistent patient IDs
Previously we generated synthetic sequential IDs for each extract but
this made it impossible to match patients across extracts which we
sometimes need to do. We now use just part of the original hash:

> EMIS patient IDs are 128 byte strings which significantly bloat our output
> files. However, because they are actually hex-encoded SHA-512 hashes we can
> get away with throwing away all but the first 63 bits and storing them as
> int64s (63 bits so we can safely store them as signed int64s without
> worrying about negative patient IDs, which might upset expectations
> elsewhere). 63 bits still gives us plenty of entropy as sqrt(2^63) is about
> 2 billion so well above the threshold where we'd need to worry about
> collisions with a population of a few tens of millions.

Note that this will produce larger CSVs than previously as patient IDs
will be up to 19 digits long whereas previously they would be at
most 8 digits. But this is still better than the 128 characters they
were initially. And the difference will disappear when using binary
output formats.