fix: Extend "prefix matching" to primary diagnosis field
When querying the Diagnoses field in the APCS data we use prefix
matching so that searching for the code `J12` will also return the code
`J120`, `J121` etc.

I initially thought this was a bug but there turns out to be a test for
precisely this behaviour. It also makes clinical sense given the
hierarchical nature of ICD-10 codes.

But we weren't doing this for the Primary Diagnosis field which is a
confusing inconsistency. It's particularly problematic with the APCS data
which seems to represent three characters codes by padding them with an
X to make them four character. As the resulting codes were not valid
ICD-10 codes there was no way to represent them in the codelist builder
and therefore. Using prefix matching sidesteps this problem.

This is possibly a slightly controversial approach; see in particular
the discussion at:
https://github.com/opensafely-core/cohort-extractor/issues/512#issuecomment-812027119

Closes #512