feat: Total of suppressed numbers in measures should be more than the threshold
If we just suppress small numbers, there is a possible
reidentification attack involving comparing the unsuppressed total to
the population size to work out the suppressed values. So we suppress
additional values over the threshold in order to bring the total
suppressed over the threshold. In practice that means suppressing
exactly one extra value (because it must itself be over the threshold
and therefore it brings the total over the threshold); to minimize the
impact on the results, we pick the smallest such value.

However this opens another hole. Where there are more than one equal
smallest value over the threshold, it may be possible for an attacker
to reorder the rows to get different entries suppressed -- thus
revealing which of the suppressed numbers are actually the small ones
and re-opening the original attack. Therefore we suppress _all_ such
equal numbers.