fix: Fix ICU queries when used with dynamic dates
The old SQL we used was weird in that it applied the date range filter
condition to the aggregated (min/max) date using a HAVING clause, rather than
filtering with a WHILE and then aggregating. This was introduced in
5ac9039, I suspect just because this was the most straightforward change
to make at the time not because there's any special reason to do the
query this way. This PR brings the SQL in line with the patterns we use
elsewhere and makes it work with dynamic dates.