fix: Don't cache the virtualenvs between jobs
This caching is intended to speed up the builds to avoid repeatedly
downloading and installing the same set of packages. However it is
somewhat unreliable: we sometimes get builds failing because the
virtualenv is unusable (for unknown reasons) and when that happens there
is no way to clear the cache without modifying the PR's commits.

This change removes the caching, sacrificing a bit of speed for improved
reliability.