feat: add query for EFI values to the TPP backend (#549)
Add a query for Electronic Frailty Index (EFI) values to the TPP backend.
To distinguish between clinical events, which are part of a patient's
EHR, and decision support values, which are not, we add the
with_these_decision_support_values function. This function is similar to
with_these_clinical_events, but omits parameters related to codelists,
parameters related to the "episodes" API (strap line: "Fake it 'till you
make it"), and deprecated parameters.