fix: Don't leak internal 9999-21-31 deregistration date
Currently active registrations are recorded as having an end date of
9999-12-31. This is an internal detail we shouldn't leak: we should
instead return a deregistration date of NULL for currently registered
patients. Aside from being inelegant, this date can't be represented as
a Pandas timestamp and so it causes binary exports to fail.

We already attempted to do the right thing here, but the code was buggy
and the tests inadequate.

Closes #500