chore: Restructure Docker build for better layering
Thanks to PySpark our dependencies are now pretty big. Previously, due
to the structure of our Dockerfile every code change however minor
required all dependencies to be re-installed into a new layer which
makes the update much larger for users to pull down.

This PR seperates out the installation of our dependencies from
the installation of the code which should mean that small code changes
require only small new layers to be pulled.