feat: Move from Presto to newer Starburst image
EMIS updated their version of Starburst and we need to update our Docker
image to match their version.

This is complicated by a rename of PrestoSQL to Trino. As a result, the
Starburst Presto image we were using is no longer receiving updates. The
first Starburst version to have this switch is 354-e. The image we now
need to use has a different name: `starburst-enterprise`.

See the [Starburst migration
page](https://docs.starburst.io/354-e/appendix/to-sep-354-plus.html#docker-image).