docs: Remove redundant value type from docstrings
Fixes #364.

This information is already in the function definition. In the
documentation, we show the default value as a separate column. This
means that we get duplication where we say "The default value is `None`"
and then show the default is `None`.

This is also brittle in that someone might modify a default, but forget
to update the docstring, or vice versa.