fix: Fix dummy data generation for categorised dates
We allow `categorised_as` functions to return dates, which can then be
used as inputs to other functions. For the purposes of real data
extraction they behave like dates and should be typed as such. However
for the purposes of dummy data generation they act like categoricals
(dates get a whole load of special treatment that we don't want here).

This addresses the issue at:
https://github.com/opensafely-core/cohort-extractor/pull/656#issuecomment-927282520