fix: Revert to using Presto client library
This reverts the change made in #625 from using
`presto-python-client=0.7.0` to using `trino==0.306.0`. That change
passes tests but triggers 500 errors from the database in production.

Rather than revert the entire PR (which would be a revert of a revert of
a revert) we leave almost all the other changes in place and do some
import munging to access the old client under the new name. This should
buy us some time while we identify the root cause.