ci: move to new style docker image building
This change is motivated by:

1) the docker build mysteriously breaking (fixes #701)
2) the image size being massive
3) consistency across project docker builds

It based on the same workflow in release-hatch and job-runner:

 - separate out build depencencies into separate build image
 - use virtualenv rather than system pip install
 - tighten up dependencies and reduce size
 - support an efficient way to do local development
 - add dev tooling to manage it

The size goes from 1950mb to 567mb in the process.

The docker image building is managed by docker-compose config, which
simplifies a bunch of things. Particularly, `run.sh` becomes a thinner
wrapper around `docker-compose run`

To enable a proper split between dev and prod deps, I've split the
requirements out into separate files, as per our other projects.

The docker expensive bits of the docker image will only now only need to
be rebuilt if any of the following files change: `setup.py`,
`requirements.{dev,prod}.in`, `dependencies.txt`,
`build-dependencies.txt`. It uses a bit of trick to
ensure this, so the Dockerfile needs a careful look.