fix: docker image python packaging
Using editable installs was causing a few issues where we had to
preserve the .egg-info directory in /app, and the docker build is
currently broken as a result.

Editable installs gave us two things: an installed package (with
entrypoints), and a custom filesystem import location added to the path.

We instead use PYTHONPATH for the latter, and for the former we install
a minimal version of the package without code, in order to get the
entrypoints installed correctly.