fix: clamp dates to pandas max
TPP uses the date `9999-12-31 00:00:00` to represent a max future date,
but this is too large for `pandas.Timestamp` when we download the
results.

So, we clamp this specific date to `pandas.Timestamp.max`.

I chose to be conservative about what dates we clamp, but in theory we
could clamp anything over `2262-04-11 23:47:16.854775807`. But that
would mean actually parsing the date string, which is what pandas is
supposed to do for us, and python stdlib does not have a generic parser
for. And I wasn't sure about the semantics.