fix: Fix population size for dummy data
When we generate dummy data, `with_value_from_file` and
`which_exist_in_file` in effect modify the population size from the
value given within *project.yaml* to the number of rows in the
user-supplied CSV file. However, when a variable in a study definition
includes a dependant date, then an ndarry of length "population size" is
assigned to a data frame of length "number of rows", which raises a
ValueError.

Here, we add add such a variable to a study definition to recreate the
error (thanks for reporting it, @rriefu). We then ensure that when these
functions are present in a study definition, the population size is set
to the number of rows in the user-supplied CSV file.