fix: Generate more efficient SQL for "summary" queries
This doesn't use exactly the approach described in the linked ticket,
although it uses the same basic technique of splitting the query in two
using a temporary table. Instead it follows the same pattern we use
elsewhere in the "events which occur on the same day" queries:
https://github.com/opensafely-core/cohort-extractor/blob/531e2080b43ab/cohortextractor/tpp_backend.py#L1420-L1466

If this turns out to be not performant enough we can try something
different, but we should consider updating all the uses of this pattern.

Closes #750