feat: add maintenance command
This will query the backend (TPP only currently) to see if its currently
running maintenance. It's designed to be called by job-runner so it can
enable job-runner db-maintenance mode.

The logic is based on the proposal by TPP on how they will populate an
events table.

The idea is that it will print "db-maintenance" if its in db mode. But
it also prints the SQL logs, which I cannot seem to supress, so
job-runner will check if the last line is "db-maintenance"

The command needs the current state passed in, or else we don't have
enough information, given TPPs proposed events.