Add urban_rural_classification mapping to docs
This is the mapping for how this variable is encoded in TPP.

I'm unsure if this is the right place for this to be documented. While the classification is commonly used elsewhere, it's possible other (future) data sources might encode it differently. On the other hand, we need it to be somewhere discoverable to people using the variable.