feat: support dummy data for generate_codelist_report command
This command does not use study definitions, so we use previously
developed code to embed a default dummy data model for it. This allows
it to be used in local development. Currently, it can only be
run in production.

Note: this work uncovered a bug in the db fixutre cleanup, in that
removing the tables didn't actually work once there was multiple tests
in the module. Rather than figure out the specific deletion order
needed, I just reused the cleanup from the tpp tests.