Fix test_date_window_behaviour_* tests
Because of this "helpful" hack [0] to increase the variety of our test
data, the bare dates were being converted to datetimes with non-zero
time parts.

This means that in test_date_window_behaviour_literals, the expected
number of events without the fix was wrong.  This was not discovered,
because that test was not run without the fix.

[0] https://github.com/opensafely-core/cohort-extractor/blob/76eb09fd43d1bd42133dd537524d2120a0d5ac66/tests/tpp_backend_setup.py#L28-L41