feat: Add support for `--output-file` argument
This allows specifying the full path to the output file, including the
output directory and extension. We need this in order to support
parameterising study definitions because at the moment the output file
name is determined by the study definition name, but we want to use the
same study definition to produce different output files.

We could acheive the same thing by adding an `--output-name` argument
and combining these with the pre-existing `--output-directory` and
`--output-format` arguments. But it seems much simpler to specify all
three together in a single argument.