fix: Allow both `--output-file` and `--output-dir` to be set
The latter is redundant, but job-runner insists on adding it to the
`run` command in any case so we can't just reject it outright. Instead
we now check that it agrees with the directory in `--output-file` and
only throw an error if they're different.

For consistency we do the same with `--output-format` although
job-runner doesn't inject that for us.