fix: Handle another type of transient database error
We've known about this type of error for ages (see [issue][1]) but when
we first wrote this error handling code (#791) we hadn't seen it in a
while so we decided to exclude it and see if it reoccurs. It has
reoccurred; not terribly often but enough times that it's worth handling
it here.

Note that transient errors which occur when downloading results are
automatically handled by the retry code so we never see them. This error
handling concerns errors which occur before we've finished generating
the results.

[1]: https://github.com/opensafely-core/cohort-extractor/issues/776#issuecomment-1091675829