feat: Add date_type parameter to patients.with_gp_consultations
There are several date columns on the Appointments table in TPP.  Until
now, we have only exposed the SeenDate column.  That is, only this
column was used for filtering, and the value in this column was returned
when the `returning` parameter was "date".

We have now added support for filtering on and returning the BookedDate
and StartDate.

Note that we do not support filtering on one column and returning the
value of another, so eg you cannot find the SeenDate of all appointments
with a BookedDate in a given range.  This would be possible to do.

Addresses part of #885.