Use VmpMapping table in temporary database, if available
As far as I can tell, the temporary database is always available in
production, but not in tests unless explicitly set via some
monkeypatching.

This means that the tests do not test the production code path, but
without significant effort I'm not sure how to do this.  However, I'm
reasonably sure that if the code is incorrect, failure will be immediate
and obvious, and we can back out this PR.