feat: Add `update_custom_medication_dictionary`
This sub-command updates the `CustomMedicationDictionary` table. (In
reality, it first drops and then creates it, but it's called "update"
for consistency with `update_vmp_mapping`.) This table is required by
opensafely-core/ehrql#1270. Whilst its maintenance should, arguably, be
handled by a separate tool, we already drop and create tables in the
temporary database in cohort-extractor, so the maintenance of this table
is here for consistency.

The implementation follows that of `cohortextractor.update_vmp_mapping`
closely. I've added a couple of QOL improvements to the test and, to
preserve my sanity, avoided using `unittest.mock`.