Remove transaction from temporary table creation
We want to ensure that the table creation and population queries happen
in separate transaction so as to avoid the locking behaviour described
in #965.

This does introduce a potential new failure mode where a job gets
halfway through populating a table and then fails, leaving the table
present but empty. A subsequent identical job will then pick up this
empty table, assume it contains the results, and attempt to download it.

In practice I don't think this will happen, and if it does it will lead
to a big obvious failure (i.e. an empty result set) rather than anything
subtle so I don't propose to do anything about it.