fix: Explicitly commit after creating temporary tables
We removed these as part of #968. However pymssql defaults to holding an
open transaction which defeats the point of the lock avoidance behaviour
we implemented so we now make sure to explicitly commit afer each step.

Theoretically we could just put the connection into autocommit mode, but
we want to do the smallest possible change here while we observe the
effect.